Add RefFnOce, RefFnMut and RefFn
